+++ src/components/Modal/Modal.js
@@ -9,6 +9,7 @@
  * </Parent>
  */
 import React, { Component } from 'react';
+import ReactDOM from 'react-dom';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import { FormattedMessage, intlShape, injectIntl } from '../../util/reactIntl';
@@ -18,23 +19,65 @@ import css from './Modal.css';
 
 const KEY_CODE_ESCAPE = 27;
 
+class Portal extends React.Component {
+  constructor(props) {
+    super(props);
+    this.el = document.createElement('div');
+  }
+
+  componentDidMount() {
+    // The portal element is inserted in the DOM tree after
+    // the Modal's children are mounted, meaning that children
+    // will be mounted on a detached DOM node. If a child
+    // component requires to be attached to the DOM tree
+    // immediately when mounted, for example to measure a
+    // DOM node, or uses 'autoFocus' in a descendant, add
+    // state to Modal and only render the children when Modal
+    // is inserted in the DOM tree.
+    this.props.portalRoot.appendChild(this.el);
+  }
+
+  componentWillUnmount() {
+    this.props.portalRoot.removeChild(this.el);
+  }
+
+  render() {
+    return ReactDOM.createPortal(this.props.children, this.el);
+  }
+}
+
 export class ModalComponent extends Component {
   constructor(props) {
     super(props);
     this.handleBodyKeyUp = this.handleBodyKeyUp.bind(this);
     this.handleClose = this.handleClose.bind(this);
+
+    this.refDiv = React.createRef();
+
+    this.state = {
+      portalRoot: null,
+    };
   }
 
   componentDidMount() {
     const { id, isOpen, onManageDisableScrolling } = this.props;
     onManageDisableScrolling(id, isOpen);
     document.body.addEventListener('keyup', this.handleBodyKeyUp);
+    this.setState({
+      portalRoot: document.getElementById('portal-root'),
+    });
   }
 
   componentDidUpdate(prevProps) {
     const { id, isOpen, onManageDisableScrolling } = prevProps;
     if (this.props.isOpen !== isOpen) {
       onManageDisableScrolling(id, this.props.isOpen);
+
+      // Because we are using portal,
+      // we need to set the focus inside Modal manually
+      if (this.props.usePorta && this.props.isOpen) {
+        this.refDiv.current.focus();
+      }
     }
   }
 
@@ -69,6 +112,7 @@ export class ModalComponent extends Component {
       intl,
       isClosedClassName,
       isOpen,
+      usePortal,
     } = this.props;
 
     const closeModalMessage = intl.formatMessage({ id: 'Modal.closeModal' });
@@ -96,7 +140,15 @@ export class ModalComponent extends Component {
     const classes = classNames(modalClass, className);
     const scrollLayerClasses = scrollLayerClassName || css.scrollLayer;
     const containerClasses = containerClassName || css.container;
-    return (
+    const portalRoot = this.state.portalRoot;
+
+    // If you want to use Portal https://reactjs.org/docs/portals.html
+    // you need to use 'userPortal' flag.
+    // ModalInMobile component needs to use the old Modal without the portal
+    // because it's relying that the content is rendered inside
+    // the DOM hierarchy of the parent component unlike Modal inside Portal.
+
+    return !usePortal ? (
       <div className={classes}>
         <div className={scrollLayerClasses}>
           <div className={containerClasses}>
@@ -105,7 +157,22 @@ export class ModalComponent extends Component {
           </div>
         </div>
       </div>
-    );
+    ) : portalRoot ? (
+      <Portal portalRoot={portalRoot}>
+        <div className={classes}>
+          <div className={scrollLayerClasses}>
+            <div
+              className={classNames(containerClasses, css.focusedDiv)}
+              ref={this.refDiv}
+              tabIndex="-1"
+            >
+              {closeBtn}
+              <div className={classNames(contentClassName || css.content)}>{children}</div>
+            </div>
+          </div>
+        </div>
+      </Portal>
+    ) : null;
   }
 }
 
@@ -120,6 +187,7 @@ ModalComponent.defaultProps = {
   isClosedClassName: css.isClosed,
   isOpen: false,
   onClose: null,
+  usePortal: false,
 };
 
 const { bool, func, node, string } = PropTypes;
@@ -137,6 +205,7 @@ ModalComponent.propTypes = {
   isClosedClassName: string,
   isOpen: bool,
   onClose: func.isRequired,
+  usePortal: bool,
 
   // eslint-disable-next-line react/no-unused-prop-types
   onManageDisableScrolling: func.isRequired,