+++ src/components/ModalInMobile/ModalInMobile.example.js
@@ -1,5 +1,5 @@
 /* eslint-disable no-console */
-import React, { Component } from 'react';
+import React, { useState } from 'react';
 import { Button } from '../../components';
 import ModalInMobile from './ModalInMobile';
 import css from './ModalInMobileExample.css';
@@ -9,41 +9,35 @@ const onManageDisableScrolling = (componentId, scrollingDisabled = true) => {
   console.log('Toggling ModalInMobile - currently:', componentId, scrollingDisabled);
 };
 
-class ModalInMobileWrapper extends Component {
-  constructor(props) {
-    super(props);
-    this.state = { isOpen: false };
-    this.handleOpen = this.handleOpen.bind(this);
-  }
+const ModalInMobileWrapper = props => {
+  const [isOpen, setOpen] = useState(false);
 
-  handleOpen() {
-    this.setState({ isOpen: true });
-  }
+  const handleOpen = () => {
+    setOpen(true);
+  };
 
-  render() {
-    return (
-      <div>
-        <div style={{ margin: '1rem' }}>Wrapper text before ModalInMobile</div>
-        <ModalInMobile
-          {...this.props}
-          onClose={() => {
-            this.setState({ isOpen: false });
-            console.log('Closing modal');
-          }}
-          isModalOpenOnMobile={this.state.isOpen}
-          onManageDisableScrolling={onManageDisableScrolling}
-        >
-          <div style={{ margin: '1rem' }}>Some content inside ModalInMobile component</div>
-        </ModalInMobile>
-        <div style={{ margin: '1rem' }}>
-          <Button onClick={this.handleOpen} className={css.visibleOnMobileLayout}>
-            Open
-          </Button>
-        </div>
+  return (
+    <div>
+      <div style={{ margin: '1rem' }}>Wrapper text before ModalInMobile</div>
+      <ModalInMobile
+        {...props}
+        onClose={() => {
+          setOpen(false);
+          console.log('Closing modal');
+        }}
+        isModalOpenOnMobile={isOpen}
+        onManageDisableScrolling={onManageDisableScrolling}
+      >
+        <div style={{ margin: '1rem' }}>Some content inside ModalInMobile component</div>
+      </ModalInMobile>
+      <div style={{ margin: '1rem' }}>
+        <Button onClick={handleOpen} className={css.visibleOnMobileLayout}>
+          Open
+        </Button>
       </div>
-    );
-  }
-}
+    </div>
+  );
+};
 
 export const Empty = {
   component: ModalInMobileWrapper,