+++ src/routeConfiguration.js
@@ -43,6 +43,12 @@ const draftSlug = 'draft';
 
 const RedirectToLandingPage = () => <NamedRedirect name="LandingPage" />;
 
+// NOTE: Most server-side endpoints are prefixed with /api. Requests to those
+// endpoints are indended to be handled in the server instead of the browser and
+// they will not render the application. So remember to avoid routes starting
+// with /api and if you encounter clashing routes see server/index.js if there's
+// a conflicting route defined there.
+
 // Our routes are exact by default.
 // See behaviour from Routes.js where Route is created.
 const routeConfiguration = () => {